requirements.txt
tqdm 
geopandas 
pandas 
ipykernel
py360convert 
pillow 
numpy
mapillary
iopaint